Move content in data dir
